# Conflicts:
#	src/main/java/com/solution/recipetalk/controller/ingredient/userhas/UserHasIngredientController.java
#	src/main/java/com/solution/recipetalk/dto/product/request/ProductRequestRegisterDTO.java
#	src/main/java/com/solution/recipetalk/service/product/impl/FindProductServiceImpl.java

+++ src/main/java/com/solution/recipetalk/service/ingredient/userhas/impl/RemoveUserHasIngredientServiceImpl.java
@@ -2,12 +2,16 @@ package com.solution.recipetalk.service.ingredient.userhas.impl;
 
 import com.solution.recipetalk.domain.ingredient.userhas.entity.UserHasIngredient;
 import com.solution.recipetalk.domain.ingredient.userhas.repository.UserHasIngredientRepository;
+import com.solution.recipetalk.exception.common.NotAuthorizedException;
 import com.solution.recipetalk.exception.ingredient.IngredientNotFoundException;
 import com.solution.recipetalk.service.ingredient.userhas.RemoveUserHasIngredientService;
+import com.solution.recipetalk.util.ContextHolder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -16,7 +20,11 @@ public class RemoveUserHasIngredientServiceImpl implements RemoveUserHasIngredie
     private final UserHasIngredientRepository userHasIngredientRepository;
     @Override
     public ResponseEntity<?> removeUserHasIngredient(Long userHasIngredientId) {
+        Long loginUserId = ContextHolder.getUserLoginId();
         UserHasIngredient byId = userHasIngredientRepository.findById(userHasIngredientId).orElseThrow(IngredientNotFoundException::new);
+        if (!Objects.equals(loginUserId, byId.getUser().getId())){
+            throw new NotAuthorizedException();
+        }
         userHasIngredientRepository.delete(byId);
         return ResponseEntity.ok(null);
     }
